migrated auth logic to controller folder and refactored
