+++ backend/routes/auth.js
@@ -1,84 +1,10 @@
 const router = require("express").Router();
-const jwt = require("jsonwebtoken");
-const User = require("../models/user.js");
+const { register, login } = require("../controllers/auth");
 
 // Register route
-router.post("/register", async (req, res) => {
-  try {
-    const { name, email, password, phoneNumber } = req.body;
-
-    // Validate GIU student email
-    if (!email.endsWith("@student.giu-uni.de")) {
-      return res.status(400).json({
-        message: "Please use your GIU student email (@student.giu-uni.de)",
-      });
-    }
-
-    // Check if user exists
-    const userExists = await User.findOne({ email });
-    if (userExists) {
-      return res.status(400).json({ message: "User already exists" });
-    }
-
-    // Create new user
-    const user = new User({
-      name,
-      email,
-      password,
-      phoneNumber,
-    });
-
-    await user.save();
-
-    // Generate token
-    const token = jwt.sign({ userId: user._id }, process.env.JWT_SECRET, { expiresIn: "7d" });
-
-    res.status(201).json({
-      token,
-      user: {
-        id: user._id,
-        name: user.name,
-        email: user.email,
-        phoneNumber: user.phoneNumber,
-      },
-    });
-  } catch (error) {
-    res.status(500).json({ message: "Server error" });
-  }
-});
+router.post("/register", register);
 
 // Login route
-router.post("/login", async (req, res) => {
-  try {
-    const { email, password } = req.body;
-
-    // Find user
-    const user = await User.findOne({ email });
-    if (!user) {
-      return res.status(400).json({ message: "Invalid credentials" });
-    }
-
-    // Check password
-    const isMatch = await user.comparePassword(password);
-    if (!isMatch) {
-      return res.status(400).json({ message: "Invalid credentials" });
-    }
-
-    // Generate token
-    const token = jwt.sign({ userId: user._id }, process.env.JWT_SECRET, { expiresIn: "7d" });
-
-    res.json({
-      token,
-      user: {
-        id: user._id,
-        name: user.name,
-        email: user.email,
-        phoneNumber: user.phoneNumber,
-      },
-    });
-  } catch (error) {
-    res.status(500).json({ message: "Server error" });
-  }
-});
+router.post("/login", login);
 
 module.exports = router;
